pyproject.toml
[project]
name = "learn-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
